requirements.txt
streamlit
openai
elevenlabs
python-dotenv
beautifulsoup4
markdown
python-docx
pandas

